refactor: move shared payment helpers up into `UnzerAbstract`
Building the basket lived in `UnzerKlarna`, fetching the customer and the risk
data in `UnzerPaylaterInvoice` -- although neither is specific to those methods.
Every Unzer provider can need them, and the next Pay later method would have had
to inherit from a sibling or copy them.

Pure move: `get_basket_id`, `build_basket_items`, `build_article_item`,
`build_shipping_item`, `build_discount_item` and `get_shipping_vat_percentage`
together with the basket item constants, plus `get_customer` and
`get_risk_data`. Behaviour is unchanged, the subclasses inherit them now.
